Java DFS: Enabled A7 DVFS
[Defect / Enhancement / New Feature]
Defect

[Problem]
Enable A7 DVFS

[Solution]
A7 lock up was observed while switching from
A7 PLL to system PLL (HWJAVA-218)

Root cause: A7 PLL was getting powered down
before the switch is complete.

Workaround: Overridden proc_ccu_set_freq_policy
function to disable PLL auto power down before
freq change and re-enable PLL auto power down once
freq switch is complete.

Increased all SS part voltage table entries 40mV

[Reviewers]

[Testing]
    [Test Cases]

    [Label/CL# Synced To]

    [Platform]
Java

[BROADCOM INTERNAL NOTES]

Reviewed-on: http://mps-gerrit.broadcom.com/22983
Branch-Open: Branch Status <branch_status_noreply@broadcom.com>
Tested-by: AutoSubmit Status <autosubmit_status_noreply@broadcom.com>

+++ drivers/regulator/bcmpmu-rgltr-bcm59xxx.c
@@ -354,7 +354,7 @@ int bcmpmu59xxx_trim_table[] = {
 
 int bcmpmu_rgltr_get_volt_id(u32 voltage)
 {
-	u8 first, last, middle, size;
+	u32 first, last, middle, size;
 	u32 volt_uv;
 
 	size = ARRAY_SIZE(bcmpmu_csr_v_table);
